app/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
sqlalchemy
psycopg2-binary
python-dotenv
requests